move invite model into the invite package and delete the old model package
