--- app/model/__init__.py
@@ -1,2 +1,0 @@
-# flake8:noqa
-from app.model.invite import Invite
